requirements.txt
python-dotenv==0.19.0
psycopg2==2.9.2
pika==1.2.0